refactor: split the author-time guard subsystem out of feature_chain_parser
feature_chain_parser.py mixed the runtime name-parsing and matching path with
a class-definition-time validation and guard-installation subsystem that only
used the class as a namespace. The author-time cluster now lives in
feature_chain_author_guards.py as module-level functions.

The import direction is one-way (the guards import the parser, never the
reverse), so there are no compatibility re-exports: the two __init_subclass__
hooks import the new module directly. A split-boundary test pins ownership,
the acyclic direction, and the absence of re-exports.

Parser: 1130 -> 691 lines.
